Implement local storage to save likes and retweets
The status of a tweet's likes/retweets is successfully saved to local storage. However, when the user refreshes the page and clicks on the like icon of a tweet that is already liked, the icon stays liked. The user must click on the icon again in order to unlike the tweet. The same behavior happens with retweets.

This behavior happens due to the way in which the code defaults the state of the likes/retweets to false.